又又修改了自动模式， 添加了双下使电机停止，增加了前端的PC控制

+++ application/cmd/robot_cmd.h
@@ -44,5 +44,5 @@ void RobotCMDInit();
 void RobotCMDTask();
 
 void Init_Value();
-float Limit_Set(float obj,int max,int min);
+float Limit_Set(float obj, float max, float min);
 #endif // !ROBOT_CMD_H
